refactor: Move `tournament.scoring_rules` to hook
